add middleware

+++ routes/web.php
@@ -44,12 +44,12 @@ Route::controller(UserController::class)->group(function () {
     Route::get('/admin-logout', 'logout');
     Route::get('/dashboard', 'dashboard')->name('dashboard');
 
-    Route::get('/receptionist', 'receptionist')->name('receptionist');
-    Route::get('/receptionist/create', 'createReceptionist')->name('create_receptionist');
-    Route::post('/receptionist/create', 'storeReceptionist')->name('store_receptionist');
-    Route::get('/receptionist/edit/{id}', 'editReceptionist')->name('edit_receptionist');
-    Route::put('/receptionist/create/{id}', 'updateReceptionist')->name('update_receptionist');
-    Route::delete('/receptionist/delete/{id}', 'deleteReceptionist')->name('delete_receptionist');
+    Route::get('/receptionist', 'receptionist')->name('receptionist')->middleware('admin');
+    Route::get('/receptionist/create', 'createReceptionist')->name('create_receptionist')->middleware('admin');
+    Route::post('/receptionist/create', 'storeReceptionist')->name('store_receptionist')->middleware('admin');
+    Route::get('/receptionist/edit/{id}', 'editReceptionist')->name('edit_receptionist')->middleware('admin');
+    Route::put('/receptionist/create/{id}', 'updateReceptionist')->name('update_receptionist')->middleware('admin');
+    Route::delete('/receptionist/delete/{id}', 'deleteReceptionist')->name('delete_receptionist')->middleware('admin');
 });
 
 Route::get('/admin', function () {
@@ -57,48 +57,48 @@ Route::get('/admin', function () {
 });
 
 Route::controller(RoomSpecController::class)->group(function () {
-    Route::get('/room_spec', 'index')->name('room_spec');
-    Route::get('/room_spec/create', 'view_create')->name('create_room_spec');
-    Route::post('/room_spec/create', 'store')->name('save_room_spec');
-    Route::get('/room_spec/edit/{id}', 'edit')->name('edit_room_spec');
-    Route::put('/room_spec/edit/{id}', 'update');
-    Route::delete('/room_spec/delete/{id}', 'destroy')->name('delete_room_spec');
+    Route::get('/room_spec', 'index')->name('room_spec')->middleware('admin');
+    Route::get('/room_spec/create', 'view_create')->name('create_room_spec')->middleware('admin');
+    Route::post('/room_spec/create', 'store')->name('save_room_spec')->middleware('admin');
+    Route::get('/room_spec/edit/{id}', 'edit')->name('edit_room_spec')->middleware('admin');
+    Route::put('/room_spec/edit/{id}', 'update')->middleware('admin');
+    Route::delete('/room_spec/delete/{id}', 'destroy')->name('delete_room_spec')->middleware('admin');
 });
 
 Route::controller(RoomController::class)->group(function () {
-    Route::get('/room', 'index')->name('room');
-    Route::get('/room/create', 'create')->name('create_room');
-    Route::post('/room/create', 'store')->name('save_room');
-    Route::get('/room/edit/{id}', 'edit')->name('edit_room');
-    Route::put('/room/edit/{id}', 'update');
-    Route::delete('/room/delete/{id}', 'destroy')->name('delete_room');
+    Route::get('/room', 'index')->name('room')->middleware('admin');
+    Route::get('/room/create', 'create')->name('create_room')->middleware('admin');
+    Route::post('/room/create', 'store')->name('save_room')->middleware('admin');
+    Route::get('/room/edit/{id}', 'edit')->name('edit_room')->middleware('admin');
+    Route::put('/room/edit/{id}', 'update')->middleware('admin');
+    Route::delete('/room/delete/{id}', 'destroy')->name('delete_room')->middleware('admin');
 });
 
 Route::controller(RoomFacilityController::class)->group(function () {
-    Route::get('/room_facility', 'index')->name('room_facility');
-    Route::get('/room_facility/create', 'create')->name('create_room_facility');
-    Route::post('/room_facility/create', 'store')->name('save_room_facility');
-    Route::get('/room_facility/edit/{id}', 'edit')->name('edit_room_facility');
-    Route::put('/room_facility/edit/{id}', 'update');
-    Route::delete('/room_facility/delete/{id}', 'destroy')->name('delete_room_facility');
+    Route::get('/room_facility', 'index')->name('room_facility')->middleware('admin');
+    Route::get('/room_facility/create', 'create')->name('create_room_facility')->middleware('admin');
+    Route::post('/room_facility/create', 'store')->name('save_room_facility')->middleware('admin');
+    Route::get('/room_facility/edit/{id}', 'edit')->name('edit_room_facility')->middleware('admin');
+    Route::put('/room_facility/edit/{id}', 'update')->middleware('admin');
+    Route::delete('/room_facility/delete/{id}', 'destroy')->name('delete_room_facility')->middleware('admin');
 });
 
 Route::controller(HotelFacilityController::class)->group(function () {
-    Route::get('/hotel_facility', 'index')->name('hotel_facility');
-    Route::get('/hotel_facility/create', 'create')->name('create_hotel_facility');
-    Route::post('/hotel_facility/create', 'store')->name('save_hotel_facility');
-    Route::get('/hotel_facility/edit/{id}', 'edit')->name('edit_hotel_facility');
-    Route::put('/hotel_facility/edit/{id}', 'update');
-    Route::delete('/hotel_facility/delete/{id}', 'destroy')->name('delete_hotel_facility');
+    Route::get('/hotel_facility', 'index')->name('hotel_facility')->middleware('admin');
+    Route::get('/hotel_facility/create', 'create')->name('create_hotel_facility')->middleware('admin');
+    Route::post('/hotel_facility/create', 'store')->name('save_hotel_facility')->middleware('admin');
+    Route::get('/hotel_facility/edit/{id}', 'edit')->name('edit_hotel_facility')->middleware('admin');
+    Route::put('/hotel_facility/edit/{id}', 'update')->middleware('admin');
+    Route::delete('/hotel_facility/delete/{id}', 'destroy')->name('delete_hotel_facility')->middleware('admin');
 });
 
 Route::controller(ReservationController::class)->group(function () {
-    Route::get('/reservation', 'index')->name('reservation');
-    Route::get('/reservation/create', 'create')->name('create_reservation');
-    Route::post('/reservation/create', 'store')->name('save_reservation');
-    Route::get('/reservation/{id}', 'detail')->name('detail_reservation');
-    Route::put('/reservation/edit/{id}', 'update')->name('update_reservation');
-    Route::delete('/reservation/delete/{id}', 'destroy')->name('delete_reservation');
+    Route::get('/reservation', 'index')->name('reservation')->middleware('receptionist');
+    Route::get('/reservation/create', 'create')->name('create_reservation')->middleware('receptionist');
+    Route::post('/reservation/create', 'store')->name('save_reservation')->middleware('receptionist');
+    Route::get('/reservation/{id}', 'detail')->name('detail_reservation')->middleware('receptionist');
+    Route::put('/reservation/edit/{id}', 'update')->name('update_reservation')->middleware('receptionist');
+    Route::delete('/reservation/delete/{id}', 'destroy')->name('delete_reservation')->middleware('receptionist');
 
     // GUEST
     Route::get('/create', 'guestCreate')->name('guestCreate');
